[TASK] Add function name to logging key
Log entries are written to an Amazon S3 bucket. The object key consists
of the date/time of the request.

This change extends the key (path) by adding the function name (e.g.
"ExtensionDetails" to it. This makes it easy to recognize for which
Amazon Lambda function a log entry has been written.