[TASK] Rework of the ExtensionDetails Lambda function
This is a complete rework of the Lambda function. It now retrieves the
extension details from the TYPO3 Extension Repository (TER) directly,
using an API call (the API was introduced in late 2020).

The information shown are formatted/styled and enriched with Markup if
possible. The author's Gravatar image is shown as a thumb URL.

Additional details about the extension and an improved error handling were
introduced. The latter outputs more meaningful error messages.